+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 [1.2.6] - 2018-11-08
 * new option: orientation
+* new option: disableWeekends
 
 [1.2.5] - 2018-09-06
 * Improve pluralize. Now as function.